PAGOPA-848 converting tests
